+++ sdk/core/Azure.Core/src/TokenCredential.cs
@@ -1,10 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Azure.Core.Pipeline;
 
 namespace Azure.Core
 {
@@ -20,6 +18,7 @@ namespace Azure.Core
         /// <param name="cancellationToken">The <see cref="CancellationToken"/> to use.</param>
         /// <returns>A valid <see cref="AccessToken"/>.</returns>
         /// <remarks>Caching and management of the lifespan for the <see cref="AccessToken"/> is considered the responsibility of the caller: each call should request a fresh token being requested.</remarks>
+        [CallerShouldAudit("https://aka.ms/azsdk/callershouldaudit/identity")]
         public abstract ValueTask<AccessToken> GetTokenAsync(TokenRequestContext requestContext, CancellationToken cancellationToken);
 
         /// <summary>
@@ -29,6 +28,7 @@ namespace Azure.Core
         /// <param name="cancellationToken">The <see cref="CancellationToken"/> to use.</param>
         /// <returns>A valid <see cref="AccessToken"/>.</returns>
         /// <remarks>Caching and management of the lifespan for the <see cref="AccessToken"/> is considered the responsibility of the caller: each call should request a fresh token being requested.</remarks>
+        [CallerShouldAudit("https://aka.ms/azsdk/callershouldaudit/identity")]
         public abstract AccessToken GetToken(TokenRequestContext requestContext, CancellationToken cancellationToken);
     }
 }