+++ sdk/openai/Azure.AI.OpenAI.Assistants/src/Custom/ThreadRun.cs
@@ -8,6 +8,11 @@ using Azure.Core;
 
 namespace Azure.AI.OpenAI.Assistants;
 
+[CodeGenSerialization(nameof(StartedAt), DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
+[CodeGenSerialization(nameof(ExpiresAt), DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
+[CodeGenSerialization(nameof(CompletedAt), DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
+[CodeGenSerialization(nameof(CancelledAt), DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
+[CodeGenSerialization(nameof(FailedAt), DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
 public partial class ThreadRun
 {
     /*
@@ -16,22 +21,6 @@ public partial class ThreadRun
      * "Required but nullable" utcDateTime doesn't emit the appropriate deserialization logic by default.
      *
      */
-
-    /// <summary> The Unix timestamp, in seconds, representing when this item was started. </summary>
-    [CodeGenMemberSerializationHooks(DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
-    public DateTimeOffset? StartedAt { get; }
-    /// <summary> The Unix timestamp, in seconds, representing when this item expires. </summary>
-    [CodeGenMemberSerializationHooks(DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
-    public DateTimeOffset? ExpiresAt { get; }
-    /// <summary> The Unix timestamp, in seconds, representing when this completed. </summary>
-    [CodeGenMemberSerializationHooks(DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
-    public DateTimeOffset? CompletedAt { get; }
-    /// <summary> The Unix timestamp, in seconds, representing when this was cancelled. </summary>
-    [CodeGenMemberSerializationHooks(DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
-    public DateTimeOffset? CancelledAt { get; }
-    /// <summary> The Unix timestamp, in seconds, representing when this failed. </summary>
-    [CodeGenMemberSerializationHooks(DeserializationValueHook = nameof(DeserializeNullableDateTimeOffset))]
-    public DateTimeOffset? FailedAt { get; }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     internal static void DeserializeNullableDateTimeOffset(