+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release History
 
+## 1.1.0-beta.2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+- Change `HttpClientPipelineTransport.Shared` from a field to a property.
+
+### Bugs Fixed
+
+### Other Changes
+
+- Made `CancellationToken` parameter passed to `BinaryContent.WriteTo` optional.
+
 ## 1.1.0-beta.1 (2024-02-01)
 
 ### Features Added